requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
atomicwrites==1.2.1       # via pytest
attrs==18.2.0             # via pytest
docutils==0.14
flake8-commas==2.0.0
flake8==3.6.0
isort==4.3.4
mccabe==0.6.1             # via flake8
more-itertools==5.0.0     # via pytest
multilint==2.4.0
patchy==1.4.0
pluggy==0.8.1             # via pytest
py==1.7.0                 # via pytest
pycodestyle==2.4.0        # via flake8
pyflakes==2.0.0           # via flake8
pygments==2.3.1
pytest-django==3.4.5
pytest==4.1.1
pytz==2018.9
pyyaml==3.13
six==1.12.0               # via more-itertools, multilint, patchy, pytest
sqlparse==0.2.4
